add user api tests
